+++ src/studio/packages/qtadmin-qtconsult/lib/consult.dart
@@ -1,7 +1,0 @@
-library qtadmin_qtconsult;
-
-export 'qtconsult.dart';
-export 'src/blocs/consult_bloc.dart';
-export 'src/constants.dart';
-export 'src/screens/qtconsult_screen.dart';
-export 'src/views/stat_item.dart';